requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Database & Vector Store
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
asyncpg==0.29.0

# GitHub Integration
PyGithub==2.1.1
PyJWT==2.8.0
cryptography==41.0.7
httpx==0.25.2

# LLM & AI
anthropic==0.7.8
langchain==0.1.0
langchain-anthropic==0.1.0
openai==1.3.7

# Vector Embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1
